Move ShaftTypeAdapter into klaxon package
